FEAT : Finish GET Method(기본 기능 구현 완료)
