+++ src/MethodProcessor/MethodProcessor.cpp
@@ -1,46 +1,191 @@
 #include "../../include/WebServ.hpp"
 
-MethodProcessor::MethodProcessor(void) {}
+//  Public
+MethodProcessor::MethodProcessor(
+    const std::vector<ServerConfigInfo>& server_list) {
+  size_t circuit = server_list.size();
+  t_entity* ret;
+  std::ifstream index;
+
+  for (size_t i = 0; i < circuit; i++) {
+    ret = new t_entity();
+    if (ret == NULL) {
+      /*error handling*/
+    }
+    ret->type_ = strdup(TYPE_HTML);
+    ret->length_ = FileSize(server_list.at(i).file_path.c_str());
+    ret->data_ = new char[ret->length_];
+    if (ret->data_ == NULL) {
+      /*error handling*/
+    }
+    index.open(server_list.at(i).file_path, std::ifstream::in);
+    index.read(ret->data_, ret->length_);
+    cache_entity_.insert(
+        std::pair<const int, t_entity*>(server_list.at(i).port, ret));
+    index.close();
+  }
+}
 
 MethodProcessor::~MethodProcessor(void) {
-  std::map<int, t_entity *>::iterator it = cache_entity_.begin();
-  while (it != cache_entity_.end()) {
-    delete[] it.operator->()->second->data_;
-    it = cache_config.PrintConfigInfos();
-    entity_.begin();
+  size_t limit = cache_entity_.size();
+  for (size_t i = 0; i < limit; i++) {
+    delete[] cache_entity_.at(i)->data_;
+    delete[] cache_entity_.at(i)->type_;
+    delete cache_entity_.at(i);
   }
   cache_entity_.clear();
 }
 
-int MethodProcessor::MethodProcessorInput(ClientMetaData *clients) {
-  struct Data *client;
+void MethodProcessor::GETFirst(int curfd, ClientMetaData* clients,
+                               std::vector<struct kevent>& change_list) {
+  Data* client = &clients->GetData();
+  switch (client->e_stage) {
+    case GET_HTML:
+      GETSecond(client);
+      break;
+    case GET_FILE:
+      GETSecondFile(client);
+      break;
+    case GET_CGI:
+      GETSecondCgi(client);
+      break;
+    case REQ_FINISHED:
+      FetchOiginalPath(client->req_message_->req_msg_.req_url_, *client);
+      if (IsCgi(client->req_message_->req_msg_.req_url_)) {
+        /*handling cgi*/
+        client->e_stage = GET_CGI;
+        int cgi_stream[2];
+        int pid = 0;
 
-  client = &clients->GetData();
-  if (client->req_message_->req_msg_.method_ == "GET") {
-    return (GET);
-  } else if (client->req_message_->req_msg_.method_ == "POST") {
-    return (POST);
-  } else
-    return (11);
+        if (pipe(cgi_stream) == -1) {
+          MakeErrorStatus(*client, 500);
+          return;
+        }
+
+        {
+          int flags = fcntl(cgi_stream[0], F_GETFL, 0);
+          fcntl(cgi_stream[0], F_SETFL, flags | O_NONBLOCK);
+          int flags = fcntl(cgi_stream[1], F_GETFL, 0);
+          fcntl(cgi_stream[1], F_SETFL, flags | O_NONBLOCK);
+        }
+
+        pid = fork();
+        if (pid == -1) {
+          // TODO : error handling
+        }
+        if (pid == 0) {
+          /* CGI handling*/
+          close(cgi_stream[0]);
+          dup2(cgi_stream[1], 1);
+
+          char** cgi_argv;
+          size_t cgi_length = client->req_message_->req_msg_.req_url_.size();
+
+          cgi_argv = new char*[2];
+          cgi_argv[0] = new char[cgi_length + 1];
+          cgi_argv[0][cgi_length] = '\0';
+          for (size_t i = 0; i < cgi_length; i += 2) {
+            cgi_argv[0][i] = client->req_message_->req_msg_.req_url_.at(i);
+            if (i + 1 < cgi_length)
+              cgi_argv[0][i + 1] =
+                  client->req_message_->req_msg_.req_url_.at(i + 1);
+          }
+          cgi_argv[1] = new char[1];
+          cgi_argv[1][0] = '\0';
+          if (execve(cgi_argv[0], cgi_argv, environ) == -1) exit(-1);
+        } else {
+          /* Parent handling*/
+          close(cgi_stream[1]);
+          ChangeEvents(change_list, cgi_stream[0], EVFILT_READ, EV_ADD, 0, 0,
+                       clients);
+          int loc = 0;
+          wait3(&loc, WNOHANG, NULL);
+          return;
+        }
+      } else if (IsFile(client->req_message_->req_msg_.req_url_, PNG) ||
+                 IsFile(client->req_message_->req_msg_.req_url_, JPG) ||
+                 IsFile(client->req_message_->req_msg_.req_url_, ICO)) {
+        /* handling File */
+        client->e_stage = GET_FILE;
+        if (IsFile(client->req_message_->req_msg_.req_url_, PNG))
+          client->entity_->type_ = strdup(TYPE_PNG);
+        else if (IsFile(client->req_message_->req_msg_.req_url_, JPG))
+          client->entity_->type_ = strdup(TYPE_JPEG);
+        else if (IsFile(client->req_message_->req_msg_.req_url_, ICO))
+          client->entity_->type_ = strdup(TYPE_ICON);
+        else
+          client->entity_->type_ = strdup(TYPE_DEFAULT);
+        int data_fd =
+            open(client->req_message_->req_msg_.req_url_.c_str(), O_RDONLY);
+        int flags = fcntl(data_fd, F_GETFL, 0);
+        fcntl(data_fd, F_SETFL, flags | O_NONBLOCK);
+        ChangeEvents(change_list, data_fd, EVFILT_READ, EV_ADD, 0, 0, clients);
+        return;
+      } else {
+        /*handling index.html*/
+        client->e_stage = GET_HTML;
+        /* cache data index.html */
+        if (cache_entity_.find(client->port_) != cache_entity_.end()) {
+          client->entity_->type_ = strdup(TYPE_HTML);
+          client->entity_->data_ = strdup(
+              cache_entity_.find(client->port_).operator*().second->data_);
+          client->entity_->length_ =
+              cache_entity_.find(client->port_).operator*().second->length_;
+          client->e_stage = GET_FINISHED;
+
+          ChangeEvents(change_list, client->event_->ident, EVFILT_WRITE,
+                       EV_ENABLE, 0, 0, NULL);
+          // TODO : kevent 들고 와야 함
+          return;
+        } else /* GET another html */ {
+          client->entity_->type_ = strdup(TYPE_HTML);
+          int data_fd =
+              open(client->req_message_->req_msg_.req_url_.c_str(), O_RDONLY);
+          int flags = fcntl(data_fd, F_GETFL, 0);
+          fcntl(data_fd, F_SETFL, flags | O_NONBLOCK);
+          ChangeEvents(change_list, data_fd, EVFILT_READ, EV_ADD, 0, 0,
+                       clients);
+          return;
+        }
+      }
+      break;
+    default:
+      /* error handling*/
+      break;
+  }
 }
 
-void MethodProcessor::MakeErrorStatus(struct Data &client, int code) {
+void MethodProcessor::POSTFirst(int curfd, ClientMetaData* clients) {
+  // TODO : POST 분기 구현
+}
+
+void MethodProcessor::DELETE(int curfd, ClientMetaData* clients) {
+  // TODO : DELETE 구현~
+}
+
+// Private
+
+void MethodProcessor::MakeErrorStatus(struct Data& client, int code) {
   client.status_code_ = code;
   if (client.entity_->data_) {
     delete[] client.entity_->data_;
   }
   client.entity_->data_ = NULL;
+  if (client.entity_->type_) {
+    delete[] client.entity_->type_;
+  }
   client.entity_->type_ = NULL;
   client.entity_->length_ = 0;
+  delete client.entity_;
 }
 
-void MethodProcessor::FetchOiginalPath(std::string &uri, struct Data &client) {
+void MethodProcessor::FetchOiginalPath(std::string& uri, struct Data& client) {
   uri.erase(0);
   uri.insert(0, client.config_->file_path);
   return;
 }
 
-bool MethodProcessor::IsFetched(std::string &uri, struct Data &client) {
+bool MethodProcessor::IsFetched(std::string& uri, struct Data& client) {
   size_t i = 0;
 
   while (client.config_->file_path[i]) {
@@ -53,8 +198,8 @@ bool MethodProcessor::IsFetched(std::string &uri, struct Data &client) {
   return (true);
 }
 
-bool MethodProcessor::IsExistFile(std::string &uri) {
-  const char *temp = uri.c_str();
+bool MethodProcessor::IsExistFile(std::string& uri) {
+  const char* temp = uri.c_str();
 
   int ret = access(temp, F_OK);
   delete[] temp;
@@ -63,7 +208,7 @@ bool MethodProcessor::IsExistFile(std::string &uri) {
   return (true);
 }
 
-bool MethodProcessor::IsCgi(std::string &uri) {
+bool MethodProcessor::IsCgi(std::string& uri) {
   std::string::reverse_iterator it = uri.rbegin();
   if (it[0] == CGI[2]) {
     if (it[1] == CGI[1]) {
@@ -75,22 +220,18 @@ bool MethodProcessor::IsCgi(std::string &uri) {
   return (false);
 }
 
-bool MethodProcessor::IsFile(std::string &uri, const char *identifier) {
-  std::string::reverse_iterator it = uri.rbegin();
+bool MethodProcessor::IsFile(std::string& uri, const char* identifier) {
+  const char* temp = uri.c_str();
 
-  size_t len = strlen(identifier);
+  int ret = access(temp, F_OK);
+  delete[] temp;
 
-  for (size_t i = 0; i < len; i++) {
-    if (it[i] == identifier[len - i])
-      continue;
-    else
-      return (false);
-  }
+  if (ret != 0) return (false);
   return (true);
 }
 
-char *MethodProcessor::CopyCstr(const char *cstr, size_t length) {
-  char *ret = new char[length + 1];
+char* MethodProcessor::CopyCstr(const char* cstr, size_t length) {
+  char* ret = new char[length + 1];
   if (ret == NULL) {
     /*error handling*/;
   }
@@ -102,105 +243,59 @@ char *MethodProcessor::CopyCstr(const char *cstr, size_t length) {
   return ret;
 }
 
-void MethodProcessor::GETSecondCgi(struct Data *client) {
-  int cgi_stream[2];
-  int pid = 0;
-
-  if (pipe(cgi_stream) == -1) {
-    MakeErrorStatus(*client, 500);
-    return;
+void MethodProcessor::GETSecondCgi(int curfd, ClientMetaData* clients,
+                                   struct kevent* cur_event) {
+  // TODO : CGI 읽기 구현
+  Data* client = &clients->GetData();
+  client->e_stage = GET_FINISHED;
+  client->entity_->length_ = cur_event->data;
+  client->entity_->data_ = new char[client->entity_->length_];
+  if (client->entity_->data_ == NULL) {
+    // TODO : error handling;
   }
-  pid = fork();
-  if (pid == -1) {
-    // TODO : error handling
-  }
-  if (pid == 0) {
-    /* CGI handling*/
-    close(cgi_stream[0]);
-    dup2(cgi_stream[1], 1);
-
-    char **cgi_argv;
-    size_t cgi_length = client->req_message_->req_msg_.req_url_.size();
-
-    cgi_argv = new char *[2];
-    cgi_argv[0] = new char[cgi_length + 1];
-    cgi_argv[0][cgi_length] = '\0';
-    for (size_t i = 0; i < cgi_length; i += 2) {
-      cgi_argv[0][i] = client->req_message_->req_msg_.req_url_.at(i);
-      if (i + 1 < cgi_length)
-        cgi_argv[0][i + 1] = client->req_message_->req_msg_.req_url_.at(i + 1);
-    }
-    cgi_argv[1] = new char[1];
-    cgi_argv[1][0] = '\0';
-    if (execve(cgi_argv[0], cgi_argv, environ) == -1) exit(-1);
-  } else {
-    /* Parent handling*/
-    close(cgi_stream[1]);
-
-    int waitloc;
-    if (wait(&waitloc) == -1) {
-      MakeErrorStatus(*client, 500);
-    } else {
-      if (WIFEXITED(waitloc)) {
-        char *buf;
-        ssize_t len = 0;
-
-        buf = new char[BUFFER_MAX + 1];
-        if (!buf) {
-          MakeErrorStatus(*client, 500);
-          return;
-        }
-        std::string temp_data;
-        while (true) {
-          len = read(cgi_stream[0], buf, BUFFER_MAX);
-          if (len == -1) {
-            delete[] buf;
-            MakeErrorStatus(*client, 500);
-            close(cgi_stream[0]);
-            return;
-          }
-          buf[BUFFER_MAX] = '\0';
-          temp_data.append(buf);
-          if (len != BUFFER_MAX) {
-            delete[] buf;
-            break;
-          }
-        }
-        size_t temp_length = temp_data.size();
-        client->entity_->length_ = temp_length;
-        client->entity_->data_ = CopyCstr(temp_data.c_str(), temp_length);
-      } else if (WIFSIGNALED(waitloc)) {
-        MakeErrorStatus(*client, 500);
-      }
-      close(cgi_stream[0]);
-      return;
-    }
-  }
-}
-
-void MethodProcessor::GETSecondFile(struct Data *client) {
-  t_entity *ret;
-  ret = new t_entity();
-  if (ret == NULL) {
-    /* error handling */
-  }
-  std::ifstream entityFile;
-
-  entityFile.open(client->req_message_->req_msg_.req_url_, std::ifstream::in);
-  entityFile.seekg(0, entityFile.end);
-  ret->length_ = entityFile.tellg();
-  entityFile.seekg(0, entityFile.beg);
-
-  ret->data_ = new char[ret->length_ + 1];
-  ret->data_[ret->length_] = '\0';
-
-  entityFile.read(ret->data_, ret->length_);
-
-  client->entity_->data_ = ret->data_;
+  int ret = read(curfd, client->entity_->data_, client->entity_->length_);
   return;
 }
 
-int MethodProcessor::FileSize(const char *filepath) {
+void MethodProcessor::GETSecondFile(int curfd, ClientMetaData* clients) {
+  // TODO : FILE 읽기 구현
+  Data* client = &clients->GetData();
+  client->e_stage = GET_FINISHED;
+  client->entity_->length_ =
+      FileSize(client->req_message_->req_msg_.req_url_.c_str());
+
+  client->entity_->data_ = new char[client->entity_->length_];
+  if (client->entity_->data_ == NULL) {
+    // TODO : error handling;
+  }
+  int ret = read(curfd, client->entity_->data_, client->entity_->length_);
+  return;
+}
+
+void MethodProcessor::GETSecond(int curfd, ClientMetaData* clients) {
+  // TODO : HTML 문서 읽기 구현
+  Data* client = &clients->GetData();
+  client->e_stage = GET_FINISHED;
+  client->entity_->length_ =
+      FileSize(client->req_message_->req_msg_.req_url_.c_str());
+
+  client->entity_->data_ = new char[client->entity_->length_];
+  if (client->entity_->data_ == NULL) {
+    // TODO : error handling;
+  }
+  int ret = read(curfd, client->entity_->data_, client->entity_->length_);
+  return;
+}
+
+void MethodProcessor::POSTSecond(int curfd, ClientMetaData* clients) {
+  // TODO : 쓰기구현
+}
+
+void MethodProcessor::POSTThird(int curfd, ClientMetaData* clients) {
+  // TODO : POST 파일 무결성 검사 구현
+}
+
+int MethodProcessor::FileSize(const char* filepath) {
   struct stat file_info;
   int ret;
 
@@ -211,80 +306,11 @@ int MethodProcessor::FileSize(const char *filepath) {
   return (file_info.st_size);
 }
 
-void MethodProcessor::GETSecond(struct Data *client) {
-  if (!IsFetched(client->req_message_->req_msg_.req_url_, *client))
-    FetchOiginalPath(client->req_message_->req_msg_.req_url_, *client);
-  if (!IsExistFile(client->req_message_->req_msg_.req_url_)) {
-    MakeErrorStatus(*client, 404);
-    return;
-  }
-  if (IsCgi(client->req_message_->req_msg_.req_url_)) {
-    GETSecondCgi(client);
-    return;
-  }
-  if (IsFile(client->req_message_->req_msg_.req_url_, PNG) ||
-      IsFile(client->req_message_->req_msg_.req_url_, JPG) ||
-      IsFile(client->req_message_->req_msg_.req_url_, ICO)) {
-    MethodGETFile(client);
-    if (IsFile(client->req_message_->req_msg_.req_url_, PNG))
-      client->entity_->type_ = strdup(TYPE_PNG);
-    else if (IsFile(client->req_message_->req_msg_.req_url_, JPG))
-      client->entity_->type_ = strdup(TYPE_JPEG);
-    else if (IsFile(client->req_message_->req_msg_.req_url_, ICO))
-      client->entity_->type_ = strdup(TYPE_ICON);
-    else
-      client->entity_->type_ = strdup(TYPE_DEFAULT);
-    return;
-  }
-  std::map<int, t_entity *>::iterator check_cache =
-      cache_entity_.find(client->port_);
-  if (check_cache != cache_entity_.end()) {
-    client->entity_->length_ = check_cache->second->length_;
-    client->entity_->data_ =
-        CopyCstr(check_cache->second->data_, client->entity_->length_);
-    client->entity_->type_ = check_cache->second->type_;
-    return;
-  }
+void ChangeEvents(std::vector<struct kevent> change_list, uintptr_t ident,
+                  int16_t filter, uint16_t flags, uint32_t fflags,
+                  intptr_t data, void* udata) {
+  struct kevent temp_event;
 
-  t_entity *ret;
-  ret = new t_entity();
-  if (ret == NULL) {
-    MakeErrorStatus(*client, 500);
-    return;
-  }
-  ret->length_ = FileSize(client->req_message_->req_msg_.req_url_.c_str());
-
-  ret->data_ = new char[ret->length_];
-  if (!ret->data_) {
-    MakeErrorStatus(*client, 500);
-    return;
-  }
-
-  std::ifstream entityFile;
-  entityFile.open(client->req_message_->req_msg_.req_url_, std::ifstream::in);
-
-  entityFile.read(ret->data_, ret->length_);
-
-  client->entity_->data_ = ret->data_;
-  client->entity_->type_ = strdup(TYPE_HTML);
-  cache_entity_.insert(std::pair<const int, t_entity *>(client->port_, ret));
-  return;
-}
-
-void MethodProcessor::GETFirst(struct Data *client) {}
-
-void MethodProcessor::POSTThird(struct Data *client) {
-  static_cast<void>(client);
-}
-
-void MethodProcessor::POSTSecond(struct Data *client) {
-  static_cast<void>(client);
-}
-
-void MethodProcessor::POSTFirst(struct Data *client) {
-  static_cast<void>(client);
-}
-
-void MethodProcessor::MethodDELETE(struct Data *client) {
-  static_cast<void>(client);
+  EV_SET(&temp_event, ident, filter, flags, fflags, data, udata);
+  change_list.push_back(temp_event);
 }
